refactor: isolate World observability transport
